chore: identify adl spec tests
