+++ package.json
@@ -108,7 +108,7 @@
     "check:unit": "mocha test/unit/",
     "check:ts": "./node_modules/typescript/bin/tsc -v && ./node_modules/typescript/bin/tsc --noEmit",
     "check:atlas": "mocha --config \"test/manual/mocharc.json\" test/manual/atlas_connectivity.test.js",
-    "check:adl": "mocha --file test/tools/runner test/manual/data_lake.test.js",
+    "check:adl": "mocha --file test/tools/runner test/manual/atlas-data-lake-testing",
     "check:aws": "mocha --file test/tools/runner test/integration/auth/mongodb_aws.test.js",
     "check:ocsp": "mocha --config \"test/manual/mocharc.json\" test/manual/ocsp_support.test.js",
     "check:kerberos": "mocha --config \"test/manual/mocharc.json\" test/manual/kerberos.test.js",
