+++ test/manual/atlas-data-lake-testing/atlas_data_lake_testing.spec.js
@@ -1,0 +1,28 @@
+'use strict';
+const path = require('path');
+const {
+  TestRunnerContext,
+  gatherTestSuites,
+  generateTopologyTests
+} = require('../../tools/spec-runner');
+
+describe('Atlas Data Lake - spec', function () {
+  const testContext = new TestRunnerContext({
+    skipPrepareDatabase: true,
+    useSessions: false,
+    user: 'mhuser',
+    password: 'pencil',
+    authSource: 'admin'
+  });
+
+  const testSuites = gatherTestSuites(
+    path.resolve(__dirname, '../../spec/atlas-data-lake-testing')
+  );
+
+  after(() => testContext.teardown());
+  before(function () {
+    return testContext.setup(this.configuration);
+  });
+
+  generateTopologyTests(testSuites, testContext);
+});
